Reorganizes tests to reflect package structure
