+++ tests/config_builder.py
@@ -1,0 +1,27 @@
+#!/usr/bin/env python3
+# -*- coding: UTF-8 -*-
+
+#    This file is part of the software package.
+#
+#    The package is free software: you can redistribute it and/or
+#    modify it under the terms of the GNU General Public License as published
+#    by the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    The package is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+#    Public License for more details.
+#
+#    You should have received a copy of the GNU General Public License
+#    along with  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+"""
+
+__authors__ = ['Daniel Perrefort']
+__copyright__ = 'Copyright 2017, Daniel Perrefort'
+
+__license__ = 'GPL V3'
+__email__ = 'djperrefort@pitt.edu'
+__status__ = 'Development'
